+++ crates/common/Cargo.toml
@@ -53,7 +53,6 @@ sov-state = { path = "../sovereign-sdk/module-system/sov-state" }
 citrea-evm = { path = "../evm" }
 citrea-primitives = { path = "../primitives/" }
 citrea-stf = { path = "../citrea-stf", features = ["native"] }
-citrea-storage-ops = { path = "../storage-ops" }
 
 [dev-dependencies]
 sov-mock-da = { path = "../sovereign-sdk/adapters/mock-da", features = ["native"] }
